Adjusted units for accuracy.

+++ cmddisplay/cmdobj.c
@@ -21,7 +21,7 @@ pol_destroy (polygon *p)
 }
 
 void
-pol_add_vertex (polygon *p, int x, int y)
+pol_add_vertex (polygon *p, double x, double y)
 {
     pos v;
     v.x = x;
