+++ cmddisplay/cmdtrf.h
@@ -3,8 +3,8 @@
 
 #include "cmdobj.h"
 
-void pol_scale      (polygon *p, int ex, int ey, int fx, int fy);
-void pol_rotate     (polygon *p, float deg, int fx, int fy);
-void pol_translate  (polygon *p, int tx, int ty);
+void pol_scale      (polygon *p, double ex, double ey, double fx, double fy);
+void pol_rotate     (polygon *p, double deg, double fx, double fy);
+void pol_translate  (polygon *p, double tx, double ty);
 
 #endif
